requirements.txt
streamlit>=1.28.0
sqlglot>=18.0.0
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.31.0
plotly>=5.0.0
pandas>=1.5.0
